pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "aiohttp>=3.11.13",
    "flask>=3.1.0",
    "lxml>=5.3.1",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.0.1",
    "telethon>=1.39.0",
    "tgcrypto>=1.2.5",
]
